Add delete functionality (including dynamic method generation) to Jeppetto
QueryModelDAO now defines deleteUsingQueryModel() and the DAOBuilder
will generate appropriate methods from either DSL style or annotations.  In
the former methods should start with "deleteBy" and in the latter the
OperationType.Delete should be specified in the @DataAccessMethod
annotation (OperationType.Read is the default).

The deleteUsingQueryModel() is initially implemented only for MongoDB.