Continue work to support optimal DBO for updates.
Update DirtyableDBObjectList to be more efficient by not tracking appended
items itself.  Instead, keeps a pointer to the point where inserts occur and
uses that to determine what items are new if the delegate doesn't need to
be rewritten.

Modify grantAccess()/revokeAccess() to not call into trueSave() as
__acl field modifications are not understood by it.  Benefit is that these
methods are now more efficient (uses $push/$pull).