Finer-grained access control options.  Support in both MongoDB and Hibernate.
One can now specify more completely which users are allowed to create, read, and
read/write objects.  To specify, one must add an @AccessControl annotation to their
model objects, specifying who can create and access these objects.

Options for creation include:
- "Identified" creators -- when the active AccessControlContext has a non-null accessId
- "Role" creators -- when the active AccessControlContext contains a specified role
- "Anonymous" creators -- for cases when no AccessControlContext is active/needed

For the "Identified" and "Role" cases, the caller is directly given a grantedAccess
(applicable independently of possibly changing values for roles).  If "None" is
specified, nothing is automatically granted and access will always depend on the
accessors option.

For access, options include:
- "Role" -- when the active AccessControlContext contains the specified role
- "Anyone" -- the object is available to anyone

For each case, what a user is allowed to do is governed by the access value.

A new set of common access control functional tests are included that are called
by both the Hibernate and MongoDB modules.