JD-1 Efficient document representation when updating an existing record.
With this change, objects that implement DirtyablyDBObject our queried to find out
which fields or values within the object have changed from when it was retrieved from
the underlying data store.  This includes all enhanced classes (via the
dirtyableDBObject.vm template), the DirtyableDBObjectList, DirtyableDBObjectMap,
and DirytableDBObjectMap.

The MongoDBQueryModelDAO will look at any object it needs to save() and determine
what the optimal update object is.  For new objects, this is just the whole thing.
For existing objects, DirtyableDBObject's getDirtyKeys() method is called and only
those items that have changed (or that we can't tell) are updated.