Add save() method that takes an explicit AccessControlContext.
In the recent update to allow Read and ReadWrite control, it is possible that the
AccessControlContext available from the AccessControlContextProvider is not appropriate to
update an existing object, and the developer would like to override it with a special value
(e.g. an "Application" context that only privileged methods can use).  In this case, we need the
equivalent of the findBy that accepts an explicit AccessControlContext, but for the save() method.
This is now available in the AccessControllable interface.

Removed 'invokeWithRole' option in DataAccessMethod since it is confusing to have two different
ways to override the AccessControlContext used during query for objects.

Implemented the above changes for both the Hibernate and MongoDB modules.  Added new tests
to validate the functionality works as expected.