Fix issue where managed Map has a key removed, then re-added.
Within DirtyableDBObjectMap, a 'removedKeys' set is maintained which is used to specify values to
$unset when the object is persisted.  If the same key is later added to the map, it will also show up
in the 'addedOrUpdatedKeys' set which means it will be part of a $set clause.  MongoDB doesn't
like it when the same field shows up in both places.  This change ensures an item is only either in
one set or the other.

Added tests to verify functionality works as expected.