Change the behavior of equals() and hashcode() for Jeppetto-managed objects
Previously, Jeppetto implemented classes' equals() and hashcode() methods.  It attempted to
determine equality based on the overall contents of an object rather than the approach the app
developer would take.  Although strictly correct, it could create a significant overhead for
equality checks, and might occasionally create false negatives (e.g. it might be too strict).

This change moves responsibility for these methods to where it belongs, the delegate
and/or super classes.  However, with this change there is one gotcha which is that the
same "object" but at different versions may be considered equal.  It will be up to the app
developer to manage this scenario  by either checking the __olv value or other fields
that will disambiguate the version such as a lastUpdatedTime).

Test cases added to validate behavior.