Support for MongoDB "view" objects.
This change lets a single MongoDB collection be used as the backing data for multiple object types
in the application.  Each DAO instance will continue to manage objects of a specific type, but if a
DAO is responsible for an object that is a view of another object, it can be configured with the
name of the class it is a "viewOf" and the "collection" where the data is found.  Jeppetto will
automatically determine what fields it should retrieve and populate the correct object.

Notes:
- Field names between the main object and view objects need to match (so Jeppetto knows what
   data to put where).
- Nested objects can be of different types (though names must match).  Jeppetto will only pull
   the requested fields for the nested objects.
- It is not required to include the "id" field in the view object.  This supports very fast data
   retrieval using MongoDB's covered index feature if the id is not part of the index.  Be
   aware that both the optimistic lock field and the access control field will be fetched if
   those features are enabled.
- View objects can be used for updates, though it requires the "id" field to be present.

Test cases added to validate behavior.