Reference-based writes: API and MongoDB implementation.
The scenario for this is that a common update needs to be made to one (or several) objects.
Previously, this would require a query to load the objects, each object would be modified, and
then each would be saved one at a time, resulting in 1+N or even N+N db calls.  With
referenced-based writes, this is now reduced to just 1 db call.  There are, of course, a couple
of caveats:

- Reference-based writes happen without reading the underlying object so it is possible
  that another thread could update the same objects.  Optimistic lock support is not yet
  available.
- Although most things can be updated using this mechanism, not every possible change
  is supported.  In particular, embedded collections cannot have items both add()ed and
  remove()d in the same update (though clear()ing and then add()ing is supported).

Besides being able to leverage the basic reference methods in GenericDAO, there is
support for DSL-style and annotation-based methods.  The DSL prefix is "referenceBy"
and works similarly to "deleteBy" methods.

Other changes include:

- Functional tests for references has been added.
- The GenericDAO now has findByIds(ID... ids) and deleteByIds(ID... ids) methods.
- The VelocityEnhancer class now takes additional context helpers that can be
  leveraged during enhancement.